Flyttet tester som ikke var knyttet til Validatoren.
